Renamed par2cmdline.h to libpar2internal.h.
In the process, removed the reference to par2cmline.h from
commandline.cpp, and fixed up the resulting breakage.
