+++ src/libpar2internal.h
