+++ src/par2repairer.cpp
@@ -18,7 +18,7 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-#include "par2cmdline.h"
+#include "libpar2internal.h"
 
 #ifdef _MSC_VER
 #ifdef _DEBUG
